fix: Fixed expectation that charge_in_kwh is available in local intelligent dispatch data

+++ custom_components/octopus_energy/api_client/intelligent_dispatches.py
@@ -11,7 +11,7 @@ class IntelligentDispatchItem:
     self,
     start: datetime,
     end: datetime,
-    charge_in_kwh: int,
+    charge_in_kwh: float,
     source: str,
     location: str
   ):
